Backend/requirements.txt
asgiref==3.7.2
Django==4.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
GDAL==3.7.2
numpy==1.25.2
Pillow==8.3.1
protobuf==4.24.3
psycopg2-binary==2.9.7
pytz==2023.3.post1
sqlparse==0.4.4
